Refactor fetchingdata functions
