+++ package.json
@@ -28,7 +28,7 @@
         "build": "react-scripts build",
         "test": "react-scripts test",
         "eject": "react-scripts eject",
-        "server": "npx nodemon Backend/server.js"
+        "server": "npx nodemon src/Backend/server.js"
     },
     "eslintConfig": {
         "extends": "react-app"
