+++ src/Frontend/Components/SettingsPage/SettingsComponents/SettingsComponents/CategoryInput.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { categories } from "../../../Data/categories";
+import { categories } from "../../../../Data/categories";
 
 function CategoryInput() {
     return (
